add initialize function to initialize mappings outside of constructor
This is a temporary hack for optimism predeploy contracts. Optimism's tools only allow to do state surgery on either constants
or primitive state variables. Mappings (or even nested mappings) and structs are not supported. Note that for predeployed contracts
the constructor isn't executed, but the state will be modified by hand.
Once nested mappings are possible to be initialized, we can move it back into the constructor. However, we may need it anyways,
once we use proxied predeploys.